seizure/requirements.txt
imbalanced-learn==0.14.1
joblib==1.5.3
numpy==2.4.2
nvidia-nccl-cu12==2.29.3
pandas==3.0.1
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sklearn-compat==0.1.5
threadpoolctl==3.6.0
xgboost==3.2.0
